Static cast to unsigned int for cases
[Retrieved from:
https://gitweb.gentoo.org/repo/gentoo.git/tree/media-libs/libmp4v2/files/libmp4v2-2.0.0-unsigned-int-cast.patch]

+++ libutil/Utility.cpp
@@ -489,7 +489,7 @@ Utility::process_impl()
         if( codes.find( code ) == codes.end() )
             continue;
 
-        switch( code ) {
+        switch( static_cast<unsigned int>( code ) ) {
             case 'z':
                 _optimize = true;
                 break;
