+++ util/mp4subtitle.cpp
@@ -164,7 +164,7 @@ SubtitleUtility::utility_option( int code, bool& handled )
 {
     handled = true;
 
-    switch( code ) {
+    switch( static_cast<unsigned int>( code ) ) {
         case LC_LIST:
             _action = &SubtitleUtility::actionList;
             break;
